Update project dependencies and remove unused files
This commit updates the package.json and package-lock.json files to reflect the latest dependencies, including the addition of Vite. It also removes several unused files, including the server.js file and various files from the node_modules directory, as well as updates the README to reflect the current tech stack. Additionally, a binary file for SQLite3 has been updated.